Add files via upload
The presentation of the analytical materials outlines the theoretical foundations of higher education strategies in the context of internationalization for the sustainable development of society. The main attention is paid to the global institutional transformations, transformative strategies of higher education, the problem of quality, personally oriented strategies, multiculturalism, communication strategies of higher education.